rename queue.py file for type checker
