+++ codex/librarian/queue_mp.py
@@ -1,4 +1,5 @@
 """Library Queue and task definitions."""
+# THIS file cannot be named queue or it causes weird type checker errors
 from dataclasses import dataclass
 from multiprocessing import SimpleQueue
